+++ tensorflow_models/models/vae.py
@@ -58,7 +58,7 @@ def create_decoder(settings, reuse=True):
 
 	with tf.variable_scope('decoder', reuse=reuse):
 		logits_x = decoder_network(settings, z_placeholder, is_training=False)
-		#dist_x_given_z = tf.contrib.distributions.Bernoulli(logits=logits_x)
+		#dist_x_given_z = tf.contrib.distributions.Bernoulli(logits=logits_x, dtype=tf.float32)
 		#decoder = tf.identity(dist_x_given_z.sample(), name='p_x_given_z/sample')
 	#return decoder
 	return tf.identity(tf.nn.sigmoid(logits_x), name='p_x_given_z/sample')
@@ -81,7 +81,7 @@ def create_probs(settings, inputs, is_training, reuse=False):
 	# Use generator to determine mean of Bernoulli distribution of reconstructed input
 	with tf.variable_scope('decoder', reuse=reuse):
 		logits_x = decoder_network(settings, z_sample, is_training=is_training)
-	dist_x_given_z = tf.contrib.distributions.Bernoulli(logits=tf_models.flatten(logits_x))
+	dist_x_given_z = tf.contrib.distributions.Bernoulli(logits=tf_models.flatten(logits_x), dtype=tf.float32)
 	
 	# NOTE: x | z is defined as over each pixel separate, where prior on z is a multivariate
 	# Hence the need to do the tf.reduce_sum op on the former to get down to a single number for each sample
@@ -97,7 +97,7 @@ def lg_likelihood(x, z, settings, reuse=True, is_training=False):
 	with tf.variable_scope('model'):
 		with tf.variable_scope('decoder', reuse=reuse):
 			logits_x = decoder_network(settings, z, is_training=is_training)
-	dist_x_given_z = tf.contrib.distributions.Bernoulli(logits=tf_models.flatten(logits_x))
+	dist_x_given_z = tf.contrib.distributions.Bernoulli(logits=tf_models.flatten(logits_x), dtype=tf.float32)
 	return tf.reduce_sum(dist_x_given_z.log_prob(tf_models.flatten(x)), 1)
 
 def lg_prior(z, settings, reuse=True, is_training=False):
